Switch to integer modulo, Update README.md
Replaced the instances of fmod(a, b) with a%b, since the operands are
both integers in all cases. Might be a bit more efficient, but easier
to follow either way!

As for the README, I updated the blog link (it wasn’t working properly
in the current file), and changed the alt texts to something more
descriptive.